made location into a modal
